mesh-1224: make ethereum outline mod
